Move non-Cabal-specific parts of Distribution.Simple.Utils out...
...into Distribution.Utils.Generic. Distribution.Simple.Utils moved items.
